PullRequest Analysis with Sonarcloud, Travis and Gradle
As the sonarqube gradle plugin is not handling the sonarqube params
from Travis, additional effort is needed to properly parse them
into the plugin.

As the variables are available as one Environment Variable, this
properties can be added to the configuration. The only Problem is
that this Environment Variable contains all the Sonarqube properties
as a JSON-Object in String representation. Hence that parsing via
JSONSlurper is needed.